requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.26
ci-info==0.3.0
click==8.1.3
clickhouse-connect==0.6.3
coloredlogs==15.0.1
configobj==5.0.8
configparser==5.3.0
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.8
duckdb==0.8.1
etelemetry==0.3.0
fastapi==0.97.0
filelock==3.12.2
fitz==0.0.1.dev2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
frontend==0.0.3
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httplib2==0.22.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.202
langchainplus-sdk==0.0.10
looseversion==1.2.0
lxml==4.9.2
lz4==4.3.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.15.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
optional-django==0.1.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pathlib==1.0.1
Pillow==9.5.0
plotly==5.15.0
posthog==3.0.1
protobuf==4.23.3
prov==2.0.0
pulsar-client==3.2.0
pydantic==1.10.9
pydot==1.4.2
PyMuPDF==1.22.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3
pyxnat==1.5
PyYAML==6.0
rdflib==6.3.2
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
tools==0.1.9
tqdm==4.65.0
traits==6.3.2
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
zstandard==0.21.0
